Fix test coverage issue with PresenterView test
The only way to ensure that the render prop is run is to fully mount and
render the component (shallow rendering skips rendering children).

The side effect was that `Nav` needed to be rendered in a tree that
contained a router, so I used `MemoryRouter` to fix that.

That, in turn, caused the Router state to be serialized with a randomly
generated key which in turn caused the snapshot test to be different
after each test run and fail. I added a `keyLength` prop which
apparently removes those auto-generated keys.

Lastly, since PresenterView is stateless, I refactored it to be a
functional component as opposed to a full class.

See also:

* https://github.com/ReactTraining/react-router/issues/4759

*
https://github.com/ReactTraining/react-router/issues/5579#issuecomment-333401692